pyproject.toml
[project]
name = "monitor360"
version = "0.1.0"
description = "Sistema profesional de monitoreo de red para ISPs"
authors = [{ name = "Juan Alvarez" }]
requires-python = ">=3.10"

dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.29",
  "python-dotenv>=1.0",
  "routeros-api>=0.18",
  "requests>=2.32",
  "pandas>=2.2",
  "supabase>=2.3",
  "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4",
  "black>=24.4",
  "isort>=5.13",
  "pre-commit>=3.7",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "responses>=0.25"
]


[tool.setuptools.packages.find]
where = ["."]
include = ["app", "monitor360*"]

[project.scripts]
monitor360-api = "monitor360.__main__:uvicorn.run"
